feat: enhance subscription management and reconciliation processes
- Added SubscriptionStatusExpired constant to models for better subscription state management.
- Implemented IsValidSubscriptionStatus function to validate subscription statuses.
- Updated reconciliation record creation to handle status updates more intelligently, preserving existing states when appropriate.
- Introduced ErrReconciliationNotRetryable to handle non-retryable reconciliation states.
- Enhanced CreateSubscription and UpdateSubscriptionStatus functions to validate subscription statuses.
- Added CreditAllApprovedUsersWithKey function to allow for operation key tracking during wallet credit operations.
- Introduced InboundIDs field in DirectPaymentProvisioningPayload for better tracking of inbound connections.
- Improved processor logic to handle completed debit transactions and ensure accurate refund amounts.
- Updated sync worker to create reconciliation records for missing subscriptions and handle database update failures more gracefully.
- Added FindClientBySubID function to XUI client for searching clients by subscription ID.
- Enhanced mergeClientConfig function to preserve unmanaged metadata during client updates.
- Added comprehensive tests for new functionality and edge cases in reconciliation and client management.

+++ internal/bot/handlers/wallet_safety.go
@@ -95,13 +95,13 @@ type SafeRefundResult struct {
 func safeRefundWalletWithDeps(
 	ctx context.Context,
 	userID int64,
-	amount float64,
+	amount int64,
 	description string,
 	originalOpKey string,
 	refundOpKey string,
 	subID *int64,
 	extra map[string]any,
-	creditFn func(ctx context.Context, userID int64, amount float64, description, operationKey string) error,
+	creditFn func(ctx context.Context, userID int64, amount int64, description, operationKey string) error,
 	persistReconFn func(ctx context.Context, record *db.ReconciliationRecord) error,
 ) SafeRefundResult {
 	if creditFn == nil {
@@ -168,7 +168,7 @@ func safeRefundWalletWithDeps(
 func safeRefundWallet(
 	ctx context.Context,
 	userID int64,
-	amount float64,
+	amount int64,
 	description string,
 	originalOpKey string,
 	refundOpKey string,
@@ -239,12 +239,12 @@ func compensateRemoteCreateDbFailure(
 	client xui.ClientConfig,
 	inboundIDs []int,
 	displayName string,
-	price float64,
+	price int64,
 	operationKey string,
 	dbErr error,
 	deleteFn func(email string) error,
 	verifyFn func(email string) (*xui.XUIClientInfo, error),
-	refundFn func(ctx context.Context, userID int64, amount float64, desc, origKey, refKey string, subID *int64, extra map[string]any) SafeRefundResult,
+	refundFn func(ctx context.Context, userID int64, amount int64, desc, origKey, refKey string, subID *int64, extra map[string]any) SafeRefundResult,
 	persistReconFn func(ctx context.Context, rec *db.ReconciliationRecord) error,
 ) RemoteCreateCompensationResult {
 	if deleteFn == nil {
